API: move lens file
